+++ Modules/Arduino/Arduino.h
@@ -12,7 +12,7 @@
 
 uint32_t millis();
 
-float constrain(float x, float a, float b);
+int32_t constrain(int32_t x, int32_t a, int32_t b);
 
 
 #endif
